+++ respy/tests/test_integration.py
@@ -1,6 +1,7 @@
 import numpy as np
 
 from respy.likelihood import get_crit_func
+from respy.tests.random_model import add_noise_to_params
 from respy.tests.random_model import generate_random_model
 from respy.tests.random_model import simulate_truncated_data
 
@@ -23,7 +24,7 @@ def test_simulation_and_estimation_with_different_models():
     df = simulate_truncated_data(params, options)
 
     # Evaluate at different points, ensuring that the simulated dataset still fits.
-    params, options = generate_random_model(point_constr=constr)
+    params = add_noise_to_params(params, perc_range=(0, 2))
     crit_func = get_crit_func(params, options, df)
 
     crit_func(params)