+++ respy/tests/test_randomness.py
@@ -3,11 +3,10 @@ import numpy as np
 import pytest
 
 import respy as rp
-from respy.config import EXAMPLE_MODELS
 from respy.tests.utils import process_model_or_seed
 
 
-@pytest.mark.parametrize("model_or_seed", EXAMPLE_MODELS + list(range(3)))
+@pytest.mark.parametrize("model_or_seed", ["kw_94_one", "kw_97_basic", None])
 def test_invariance_of_model_solution_in_solve_and_criterion_functions(model_or_seed):
     params, options = process_model_or_seed(model_or_seed)